chore: restructure data directory
