--- app/src/main/java/org/hospitalmanager/model/UpdatePassword.java
@@ -1,23 +1,0 @@
-package org.hospitalmanager.model;
-
-public class UpdatePassword {
-    public static class Request {
-        private String newPassword;
-
-        public String getNewPassword() {
-            return newPassword;
-        }
-    }
-
-    public static class Response {
-        private String message;
-
-        public Response(String message) {
-            this.message = message;
-        }
-
-        public String getMessage() {
-            return message;
-        }
-    }
-}
